separate encoder and decoder
