+++ QB3.vcxproj.filters
@@ -20,6 +20,9 @@
     <ClCompile Include="bmap.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="QB3tables.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="bmap.h">
@@ -32,6 +35,15 @@
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="bitstream.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="QB3decode.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="QB3encode.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="QB3common.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
